Backend/requirements.txt
fastapi
uvicorn
passlib[bcrypt]
python-jose
motor
python-dotenv
smtplib
